refactor(models): distribute types into owning store modules
Move all type definitions from src/models.rs into their owning modules:
- Channel, ChannelType, ChannelMember → store/channels.rs
- SenderType, Message, ReceivedMessage, AttachmentRef, HistoryMessage,
  ActivityMessage → store/messages.rs
- Task, TaskStatus, TaskInfo, ClaimResult → store/tasks.rs
- Agent, AgentEnvVar, AgentStatus, AgentConfig, Human → store/agents.rs
- KnowledgeEntry, RememberRequest/Response, RecallQuery/Response → store/knowledge.rs
- Attachment, ServerInfo, ChannelInfo, AgentInfo, HumanInfo → store/mod.rs
- ActivityEntry, ActivityLogEntry, ActivityLogResponse → activity_log.rs
- API DTOs (SendRequest, CreateAgentRequest, etc.) → server/handlers.rs

All submodules changed from private to pub mod. Update all import sites
(agent/manager.rs, server/mod.rs, server/handlers.rs, main.rs, all tests).
Zero behavior changes; all 69 tests pass.
