+++ src/lib.rs
@@ -1,6 +1,5 @@
 pub mod activity_log;
 pub mod agent;
 pub mod bridge;
-pub mod models;
 pub mod server;
 pub mod store;
