+++ tests/e2e_tests.rs
@@ -1,6 +1,7 @@
 use std::sync::Arc;
 
-use chorus::models::{ChannelType, SenderType};
+use chorus::store::channels::ChannelType;
+use chorus::store::messages::SenderType;
 use chorus::server::build_router;
 use chorus::store::Store;
